+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/NetworkManagerPeripheral.kt
@@ -2,10 +2,16 @@ package site.siredvin.peripheralworks.computercraft.peripherals
 
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
+import dan200.computercraft.api.peripheral.IComputerAccess
+import net.minecraft.core.BlockPos
+import net.minecraft.core.Direction
+import net.minecraft.world.level.block.entity.BlockEntity
+import net.minecraft.world.level.block.state.properties.BlockStateProperties
 import site.siredvin.peripheralworks.common.blockentity.NetworkManagerBlockEntity
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
 import site.siredvin.tweakium.modules.peripheral.OwnedPeripheral
 import site.siredvin.tweakium.modules.peripheral.owner.BlockEntityPeripheralOwner
+import site.siredvin.tweakium.modules.peripheral.representation.LuaRepresentation
 
 class NetworkManagerPeripheral(private val be: NetworkManagerBlockEntity) :
     OwnedPeripheral<BlockEntityPeripheralOwner<NetworkManagerBlockEntity>>(
@@ -85,5 +91,33 @@ class NetworkManagerPeripheral(private val be: NetworkManagerBlockEntity) :
         return MethodResult.of(
             *group.peripherals.filter { be.peripherals.contains(it) }.toTypedArray(),
         )
+    }
+
+    @LuaFunction(mainThread = true)
+    fun getDistanceBetween(computer: IComputerAccess, firstName: String, secondName: String): MethodResult {
+        val first = computer.getAvailablePeripheral(firstName) ?: return MethodResult.of(null, "Cannot find first peripheral")
+        val second = computer.getAvailablePeripheral(secondName) ?: return MethodResult.of(null, "Cannot find second peripheral")
+        val firstTarget = first.target
+        val secondTarget = second.target
+        val firstPos = firstTarget as? BlockPos
+            ?: if (firstTarget is BlockEntity) {
+                firstTarget.blockPos
+            } else {
+                return MethodResult.of(null, "Cannot determine first target position")
+            }
+        val secondPos = secondTarget as? BlockPos
+            ?: if (secondTarget is BlockEntity) {
+                secondTarget.blockPos
+            } else {
+                return MethodResult.of(null, "Cannot determine second target position")
+            }
+        val firstBlockState = peripheralOwner.level!!.getBlockState(firstPos)
+        val distance = BlockPos(firstPos.x - secondPos.x, firstPos.y - secondPos.y, firstPos.z - secondPos.z)
+        val facing = if (firstBlockState.properties.contains(BlockStateProperties.HORIZONTAL_FACING)) {
+            firstBlockState.getValue(BlockStateProperties.HORIZONTAL_FACING)
+        } else {
+            Direction.EAST
+        }
+        return MethodResult.of(LuaRepresentation.forBlockPos(distance, facing.opposite, BlockPos(0, 0, 0)))
     }
 }